readyset-client: Move debug info types to debug::info
Rather than jamming these types, which are all really just used for
getting debug information about the graph, into lib.rs in
readyset_client, move them to the debug::info module, which is a much
better place for them.

Other than that nothing's changed about the types themselves.
