+++ dataflow-state/src/lib.rs
@@ -20,8 +20,9 @@ use derive_more::From;
 use hashbag::HashBag;
 use itertools::Either;
 pub use partial_map::PartialMap;
+use readyset_client::debug::info::KeyCount;
 use readyset_client::internal::Index;
-use readyset_client::{KeyComparison, KeyCount};
+use readyset_client::KeyComparison;
 use readyset_data::DfValue;
 use readyset_errors::ReadySetResult;
 use replication_offset::ReplicationOffset;
