+++ dataflow-state/src/memory_state.rs
@@ -4,8 +4,9 @@ use std::rc::Rc;
 use common::{IndexType, Record, Records, SizeOf, Tag};
 use hashbag::HashBag;
 use rand::{self, Rng};
+use readyset_client::debug::info::KeyCount;
 use readyset_client::internal::Index;
-use readyset_client::{KeyComparison, KeyCount};
+use readyset_client::KeyComparison;
 use readyset_data::DfValue;
 use readyset_errors::ReadySetResult;
 use replication_offset::ReplicationOffset;
