+++ readyset-dataflow/src/domain/mod.rs
@@ -25,8 +25,9 @@ pub use internal::{DomainIndex, ReplicaAddress};
 use merging_interval_tree::IntervalTreeSet;
 use petgraph::graph::NodeIndex;
 use readyset_alloc::StdThreadBuildWrapper;
+use readyset_client::debug::info::KeyCount;
 use readyset_client::internal::Index;
-use readyset_client::{channel, internal, KeyComparison, KeyCount, ReaderAddress};
+use readyset_client::{channel, internal, KeyComparison, ReaderAddress};
 use readyset_errors::{internal, internal_err, ReadySetError, ReadySetResult};
 use readyset_util::futures::abort_on_panic;
 use readyset_util::progress::report_progress_with;
